Extract form submission logic for collection form
